requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
flask-cors==3.0.10
cryptography==41.0.7
eventlet==0.33.3
python-engineio==4.3.1
python-socketio==5.4.0
requests==2.26.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-html==3.2.0 